+++ src/main/java/com/ace/erp/service/sys/AuthService.java
@@ -2,7 +2,7 @@ package com.ace.erp.service.sys;
 
 import com.ace.erp.entity.sys.Auth;
 import com.ace.erp.entity.sys.User;
-import com.ace.erp.shiro.persistence.AuthMapper;
+import com.ace.erp.persistence.AuthMapper;
 import org.apache.commons.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -63,7 +63,7 @@ public class AuthService {
      * @param organizationId    可选
      * @return
      */
-    public String findRoleIds(Integer userId, Long organizationId) {
+    public String findRoleIds(Integer userId, Integer organizationId) {
 
         Map<String, Object> params = new HashMap<String,Object>();
         //StringBuilder hql = new StringBuilder("select roleIds from Auth where ");
